change modules to use fixpoint (like pkg overlays)
We want modules to be able to (1) change the busybox derivation
(e.g. to add applets), but (2) to all refer to the *same* busybox,
so that we don't have six different copies of it in the eventual image.
This seems to call for the "overlays" pattern, a.k.a. fixpoint

I am not 100% that this actually works as I think it does, but it seems
not to have done anything unexpected yet